+++ library/src/schemas/intersect/intersect.ts
@@ -3,8 +3,14 @@ import type {
   ErrorMessage,
   Issues,
   MaybeReadonly,
+  Pipe,
 } from '../../types/index.ts';
-import { parseResult, schemaIssue } from '../../utils/index.ts';
+import {
+  defaultArgs,
+  parseResult,
+  pipeResult,
+  schemaIssue,
+} from '../../utils/index.ts';
 import type { IntersectInput, IntersectOutput } from './types.ts';
 import { mergeOutputs } from './utils/index.ts';
 
@@ -34,25 +40,55 @@ export type IntersectSchema<
    * The error message.
    */
   message: ErrorMessage;
+  /**
+   * The validation and transformation pipeline.
+   */
+  pipe: Pipe<IntersectInput<TOptions>> | undefined;
 };
 
 /**
  * Creates an intersect schema.
  *
  * @param options The intersect options.
- * @param message The error message.
+ * @param pipe A validation and transformation pipe.
  *
  * @returns An intersect schema.
  */
 export function intersect<TOptions extends IntersectOptions>(
   options: TOptions,
-  message: ErrorMessage = 'Invalid type'
+  pipe?: Pipe<IntersectInput<TOptions>>
+): IntersectSchema<TOptions>;
+
+/**
+ * Creates an intersect schema.
+ *
+ * @param options The intersect options.
+ * @param message The error message.
+ * @param pipe A validation and transformation pipe.
+ *
+ * @returns An intersect schema.
+ */
+export function intersect<TOptions extends IntersectOptions>(
+  options: TOptions,
+  message?: ErrorMessage,
+  pipe?: Pipe<IntersectInput<TOptions>>
+): IntersectSchema<TOptions>;
+
+export function intersect<TOptions extends IntersectOptions>(
+  options: TOptions,
+  arg2?: Pipe<IntersectInput<TOptions>> | ErrorMessage,
+  arg3?: Pipe<IntersectInput<TOptions>>
 ): IntersectSchema<TOptions> {
+  // Get message and pipe argument
+  const [message = 'Invalid type', pipe] = defaultArgs(arg2, arg3);
+
+  // Create and return intersect schema
   return {
     type: 'intersect',
     async: false,
     options,
     message,
+    pipe,
     _parse(input, info) {
       // Create typed, issues, output and outputs
       let typed = true;
@@ -108,8 +144,8 @@ export function intersect<TOptions extends IntersectOptions>(
           output = result.output;
         }
 
-        // Return typed parse result
-        return parseResult(true, output, issues);
+        // Execute pipe and return typed parse result
+        return pipeResult(output, this.pipe, info, 'intersect', issues);
       }
 
       // Otherwise, return untyped parse result