+++ library/src/schemas/variant/variantAsync.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { parseAsync } from '../../methods/index.ts';
+import { custom, customAsync } from '../../validations/index.ts';
 import { boolean } from '../boolean/index.ts';
 import { literal } from '../literal/index.ts';
 import { number } from '../number/index.ts';
@@ -51,5 +52,66 @@ describe('variantAsync', () => {
         null
       )
     ).rejects.toThrowError(error);
+  });
+
+  test('should execute pipe', async () => {
+    const error = 'Input is invalid';
+
+    const schema1 = variantAsync(
+      'type',
+      [
+        object({ type: literal('a'), a: string() }),
+        object({ type: literal('b'), b: number() }),
+      ],
+      [
+        custom(
+          (input) =>
+            (input.type === 'a' && input.a === 'test') ||
+            (input.type === 'b' && input.b === 10),
+          error
+        ),
+      ]
+    );
+    const input1 = { type: 'a', a: 'test' };
+    const output1 = await parseAsync(schema1, input1);
+    expect(output1).toEqual(input1);
+    const input2 = { type: 'b', b: 10 };
+    const output2 = await parseAsync(schema1, input2);
+    expect(output2).toEqual(input2);
+    expect(() =>
+      parseAsync(schema1, { type: 'a', a: 'foo' })
+    ).rejects.toThrowError(error);
+    expect(() =>
+      parseAsync(schema1, { type: 'b', b: 123 })
+    ).rejects.toThrowError(error);
+
+    const schema2 = variantAsync(
+      'type',
+      [
+        object({ type: literal('a'), a: string() }),
+        object({ type: literal('b'), b: number() }),
+      ],
+      'Error',
+      [
+        customAsync(
+          async (input) =>
+            (input.type === 'a' && input.a === 'test') ||
+            (input.type === 'b' && input.b === 10),
+          error
+        ),
+      ]
+    );
+    const input3 = { type: 'a', a: 'test' };
+    const output3 = await parseAsync(schema2, input3);
+    expect(output3).toEqual(input3);
+    const input4 = { type: 'b', b: 10 };
+    const output4 = await parseAsync(schema2, input4);
+    expect(output4).toEqual(input4);
+    expect(parseAsync(schema2, { type: 'a', a: 'foo' })).rejects.toThrowError(
+      error
+    );
+    expect(parseAsync(schema2, { type: 'b', b: 123 })).rejects.toThrowError(
+      error
+    );
   });
 });